Rename DescribeSystemStep
